// foundation-layer/src/publishing/enums/rulebook-section-type.enum.ts
export enum RulebookSectionType {
  CANONICAL_QUESTION = "CANONICAL_QUESTION",
  CONTRACT_DEFINITION = "CONTRACT_DEFINITION",
  RESOLUTION_CRITERIA = "RESOLUTION_CRITERIA",
  SOURCE_POLICY = "SOURCE_POLICY",
  TIME_POLICY = "TIME_POLICY",
  EDGE_CASES = "EDGE_CASES",
  INVALIDATION = "INVALIDATION",
  EXAMPLES = "EXAMPLES",
}